Introduced new `IsDirty()` overloads for `AttributeValueCollection`
Previously, there was an `IsDirty()` method that accepted an attribute key (`name`) to check the status of an individual attribute. In most use cases, though, it makes more sense to check the entire `AttributeValueCollection` to see if there are _any_ dirty attributes, as that's a condition that will determine if the topic itself should be saved. To support this, a new overload is introduced which, absent any parameters, will return true of any `AttribtueValue`s stored in it are marked as `IsDirty`.

In addition, however, this presents an opportunity to centralize logic for identifying extraneous byline (`LastModifiedBy`) and dateline (`LastModified`) attributes. These attribute values are automatically and dynamically generated by some clients, such as the **OnTopic Editor**. As a result, if you save a topic five times without making any changes, you'll end up with five versions due to the `DateModified` being automatically updated. To help address that in a centralized fashion, the `IsDirty()` method also accepts an `excludeLastModified` parameter which will only report `true` if there are `AttributeValue`s marked as `IsDirty` whose `Key`s don't start with `LastModified`. This provides a useful condition for excluding extraneous updates—and, therefore, "orphaned" versions.

This supports Remove byline and dateline if no other attributes are updated (#20).