Skip indexed attributes if only the dateline, byline have changed
In the previous commit (d3b49ff), we introduced an overload to `AttributeValueCollection` which can determine if any attribute values have changed (`IsDirty`), while optionally excluding the dateline (`LastModified`) and byline (`LastModifiedBy`).

In this commit, we implement this new `IsdDirty()` method as a condition around the indexed attributes in `Save()`. This ensures that the indexed attributes aren't updated—thus bypassing the `LastModified` and `LastModifiedBy` updates—if no attributes have been updated outside of the auto-generated dateline and byline.

This is useful for preventing extraneous and otherwise unnecessary attributes and versions from being added to the database. This will also allow us to remove similar code from the **OnTopic Data Exchange** library—which deals with the same issue when importing JSON data—thus centralizing that implementation.

Technically, these extraneous fields don't hurt anything. But they tend to clutter up the database, making it a bit cumbersome to sort through attribute versions. They can also introduce a lot of versions which, if rolled back to, don't actually change any attribute values outside of the dateline and byline. As such, removing these helps reduce noise which, in most cases, isn't adding any value.

(Usually this happens because an editor saves a topic after changing nested topics—which is unnecessary, but difficult to communicate via the editor, so we've accepted it as an otherwise harmless action some editors will take.)

This will result in skipping a dateline and byline if an editor only modifies relationships. In the future, we might consider making a more sophisticated version of this which accounts for relationships, but for now that's an outside use case that adds a lot of complexity and isn't a priority.

This supports Remove byline and dateline if no other attributes are updated (#20).