Added XMLDoc documentation for `Relationships` enum
Since this is primarily used by the `FollowAttribute` and `TopicMappingService`, I focused the documentation on those use cases. It's worth noting, however, that this _could_ be used in other locations as well. Keeping the documentation focused on a general would make it less useful for its primary use case, though, so I opted for specificity rather than keeping it overly generic. If we find ourselves using it in a lot of other places, we may want to revisit this in the future.