Removed exception on empty database
Previously, an exception was thrown whenever `Load()`'s call to the `LoadTopics` stored procedure failed to return any topics. That made good sense. But it also prevents us from being able to initialize an empty database using the new **OnTopic Data Exchange** library that's now integrated with the **OnTopic Editor**.

To mitigate this, the exception is still thrown, but only when loading a _specific_ `topicId`; when trying to retrieve the entire topic graph—as is typically done when warming the cache when an application loads—it won't throw an exception. This is obviously a rare scenario that should only occur when working with an empty database. This provides an opportunity for developers to use the **OnTopic Data Exchange** library to import one or more JSON files to establish a topic schema.