Ensure content type descriptor initialization
With the introduction of the new `GetContentTypeDescriptors()` overload (7a5a243), we inadvertantly introduced the possibility of bypassing the main `GetContentTypeDescriptors()` logic entirely, thus preventing even checking to see if content types exist in the database before looking for them in the local cache.

In practice, this probably wouldn't hurt anything. But it could potentially introduce problems if working with partial or disconnected topic graphs, where the 'Configuration:ContentTypes` is not part of the in-memory topic graph.

To mitigate this, we now call `GetContentTypeDescriptors()` from the new overload. But we don't want to either a) call that multiple times, or b) introduce a circular loop. So, to address that risk, we now exclusively initialize the `_contentTypeDescriptors` cache from within `GetContentTypeDescriptors()` method, and always access the cache through that.