Fixed bug when querying root
When we introduced the `Topic.GetTopicByUniqueKey()` extension method (f2cb109), we significantly optimized the implementation relative to the previous, analogous implementation in `CachedTopicRepository.GetTopic()` (867b110). In doing so, however, we also introduced an inadvertent loophole: It fails to return the unique topic is the unique topic is the root topic.

This is easily remedied by checking this condition at the top of the `GetTopicByUniqueKey()` method. In addition, to be safe, I introduced two unit tests for this scenario. One is for the extension itself (as part of `TopicQueryingTest`). The other is for another extension method, `TopicRepositoryExtensions.Load()` (via `TopicRepositoryExtensionsTest`) which currently relies on that extension method.

(In practice, these two unit tests are evaluating the same underlying logic and are, thus, redundant. But in the past we've modified the underlying logic used by `TopicRepositoryExtensions.Load()`, and so it's useful to evaluate this at that level as well.)