Bypass unsaved relationships
Typically, relationships will point to established topic entities that have been saved. When saving large topic graphs, however, relationships may point to topics which have not yet been saved—in which case they will be pointing to a topic with the `Id` of `-1`. If this is persisted to the database, a foreign key violation will occur.

This is a known limitation of the library, which currently requires calling `Save()` twice in these scenarios. That's an unfortunate if edge scenario. But that scenario should not yield FKC violations on the initial `Save()`. To mitigate that, we'll just ignore any relationships pointing to an unsaved `Topic`.