Don't persist unsaved `Topic.DerivedTopic`
Previously, when the `Topic.DerivedTopic` was set, the underlying `Topic.Attributes`'s `TopicID` attribute was also set—even if the `DerivedTopic.Id` was `-1` (i.e., meaning it was unsaved). This provides no value to the persistence store since that `Topic.Id` cannot be used to rehydrate the topic graph.

This update prevents this from happening by putting in a check that the `DerivedTopic.Id` is a positive integer (i.e., has been saved). Of course, the source topic will need to be saved once the `DerivedTopic` is saved to ensure that the relationship is persisted.

This brings this functionality in line with how relationships work.

Includes unit tests for validating the behavior.