Reevaluate `DerivedTopic` on `Save()`
If a `DerivedTopic` is set prior to it being saved, the `DerivedTopic.Id` won't be persisted as a `TopicID` attribute in the `Topic.Attributes` collection. This is by design to prevent an invalid relationship from being persisted to the underlying data store (a3b6da1).

The problem with this, however, is that the `Topic` has no (current) way of knowing when the target `DerivedTopic` has been saved, and thus its underlying `TopicID` attribute will remain unset even after the `DerivedTopic` has received a valid `Topic.Id`. To mitigate this, the `TopicRepositoryBase.Save()` method now sets `DerivedTopic` to itself. If the `Id` has changed—i.e., if the `Id` has been set—then that will cause it to be persisted to the `Topic.Attributes` collection and subsequently persisted to the underlying data storage.

This fixes a bug that occurred when recursively saving in-memory topic graphs, as otherwise `DerivedTopic`s would only be persisted if either a) they were saved before the source class, or b) the `DerivedTopic` relationship was reestablished before a subsequent `Save()`. This is exactly how e.g. the **OnTopic Editor** works with the **OnTopic Data Transfer** library—but we shouldn't rely on implementors knowing this. By baking this into `Save()`, we help prevent unintuitive results when working with these scenarios.