Introduced new `TopicNotFoundException`
In order to provide more fine-tuned error handling of cases where a specific topic is requested from the repository, but cannot be found, we are now returning a new `TopicNotFoundException`. As this derives from the previous `TopicRepositoryException`, this should remain compatible with previous exception handlers, while offerring callers more specificity.

(Though, it's worth noting, there is a bug in `Load(string topicKey)` which trips on the retrieval of the return code from the `GetTopicId` function if no topic is found. As such, it continues to throw an off-target exception. Further, `Load(int topicKey)` previously threw a `NullReferenceException` if a topic wasn't found, which it shouldn't have been doing. Fortunately, it wasn't publicly documented, and no implementers have been catching it, so we're going to treat this as a bug fix, not a breaking change.)