Swallow `TopicNotFoundException` in `GetContentTypeDescriptors()`
When the database is empty, the call to `ITopicRepository.Load("Configuration")` is _expected_ to return a `TopicNotFoundException`. As part of the empty repository support, however, we don't want this to prevent rendering of the application when `GetContentTypeDescriptors()` is called. As such, we're swalling this exception.

Unfortunately, this is introducing a `CA1031`, which happens when capturing a general exception. The only problem is that this isn't a general exception. This used to be a bug in Roslyn Code Analyzers, but it was fixed in 2.9.7. It may have been reintroduced in 3.0.0? Regardless, I've reported the issue to the Roslyn team, but in the meanwhile we'll need to suppress the issue.