Handle potential null return codes in `GetReturnCode()`
The `GetReturnCode()` method assumes that the operation it is applied to will return a `@ReturnCode` (or a parameter corresponding to the `sqlParameter` parameter). If the operation doesn't return one, an unintuitive and unexpected exception is thrown.

This is mitigated in this commit by instead falling back to a generic error code of `-1`. This addresses, for instance, a bug where the `GetTopicId()` function wasn't returning a value if the `@TopicKey` parameter couldn't be found. That bug has since been fixed (d517971)—but we should be setup to handle that potentiality regardless. This also provides flexibility for databases that haven't updated `GetTopicId()` yet.