Throw `TopicNotFoundException` in `Load(int topicId, DateTime version)`
Implemented the new `TopicNotFoundException` (2300e69) in `Load(int topicId, DateTime version)`. This was previously implemented for `Load(int topicId)` and `Load(string topicKey)`.

These implementations previously threw a `NullReferenceException` if a topic wasn't found, which it shouldn't have been doing. Fortunately, it wasn't publicly documented, and no implementers have been catching it, so we're going to treat this as a bug fix, not a breaking change.