Introduce new `IsExtendedAttribute` property to `AttributeValue`
The object model doesn't care whether or not an `AttributeValue` was loaded from an `IsExtendedAttribute` repository store or not. Nevertheless, keeping track of this is useful for if the attribute is `Save()`d, as it allows us to detect potential mismatches between where the attribute _was_ saved, and where it is _supposed_ to be saved. As such, tracking this will enable us—in a future commit!—to detect cases where an attribute value hasn't changed (i.e., `!IsDirty`) _but_ the attribute should be saved _anyway_ since its _configured storage location_ (from `AttributeDescriptor.IsExtendedAttribute`) has changed since the attribute value was last saved.