Introduce new `isExtendedAttribute` parameter to `SetValue()`
In order for `ITopicRepository` to establish that an `AttributeValue` is an `IsExtendedAttribute`, it needs to be able to pass that value to an overload of `SetValue()`. To handle this, a new, optional `isExtendedAttribute` parameter is appended to the `SetValue()` method to allow callers to specify the `IsExtendedAttribute` property of the underlying `AttributeValue`.

This also requires updating a few XMLDocs which reference the `SetValue() method. Each of those pieces of documentation must be updated to reflect the updated signature and potentially "repaginated" to maintain our 128 character code width standard.