Enable explicit connection pooling on `Save()`
While ADO.NET's built-in connection pooling _should_ be sufficient, with the new handling of unresolved references, it's easy to maintain an open connection for the duration of the recursive save. For most cases, where the save isn't recursive, this won't impact performance one way or the other. For recursive saves of large topic graphs, however, it eliminates some overhead of opening and closing a connection for every `Topic` in the topic graph.