Simplified `skipXml` logic
It's unclear if we still need `skipXml` and we should probably reevaluate it as part of v5.0.0. For this version, however, it remains part of the interace. While we have it, we should follow Ignia's coding standards and use braces when dealing with `if`/`then` conditions. Or, better yet, just streamline it into a single line using the ternary operator.