In `PersistRelations()`, clear relationships `DataTable` each loop
This resolves a potential bug where related topics from a previous relationship might have been persisted to any subsequent relationships on the same `Topic`. This would only affect topics of a content type with multiple relationships on them.

Fortunately, in the only cases we had of that, the secondary relationships a) weren't used any longer, and b) were the last relationship alphabetically. As such, this bug shouldn't have impacted any current implementations. Phew!