Introduced `IsDirty` flag to `NamedTopicCollection`
The `IsDirty` flag tracks whether or not the collection has been modified in a meaningful way. By default, this value is set by overriding protected methods for `InsertItem()`, `SetItem()`, `RemoveItem()`, and `ClearItems()`, which covers underlying methods such as `Add()`, `Remove()`, &c. from a variety of enty points.

The `IsDirty` flag can be reset by any external agent. For instance, an `ITopicRepository` can populate the relationships, then reset `IsDirty` to `false` to indicate that this was part of the preloading process.

Attempts are made to eliminate false positives. If a `topic.Key` already exists, for instance, when calling `InsertItem()` or `SetItem()`, the `IsDirty` flag won't be enabled. In these cases, an `ArgumentException` will likely be thrown due to a duplicate key, but even if that's handled by the caller we don't want `IsDirty` to be set. Likewise, if `ClearItems()` is called, it first checks to confirm that any topics exist.

Includes a set of unit tests for evaluating various scenarios—including potential sources of false positives, such as inserting duplicates, clearing empty collections, and removing items that weren't already in the collection.