Conditionally reset `NamedTopicCollection.IsDirty` on `Save()`
When calling `PersistRelations()`, set the `NamedTopicCollection`'s `IsDirty` property (4762694) to `false` after finishing writing its values to the underlying SQL database. That way, we can check for this during `Save()` to optionally skip `PersistRelations`.

This only happens if there aren't any unresolved references—i.e., when all related topics have been saved. If there are, then `isDirty` will be set to `true`. That way, we're not at risk of subsequent calls to `Save()` bypassing `PersistRelations()`.  This prevents a potential scenario where all saved topic references are persisted, the collection is marked `!IsDirty`, and then unresolved references aren't subsequently saved because the collection isn't marked as `IsDirty`. That condition can't happen with the current code base, but it's a condition we'd _like_ to introduce in the future, and this will help keep the data accurate.