Ignore relationships if `!RelatedTopicCollection.IsDirty()`
Previously, whenever doing a `Save()`, relationships were deleted (as part of the `UpdateTopic` stored procedure) and subsequently recreated (as part of the `PersistRelations` stored procedure). This adds in a condition to only do this if the `RelatedTopicCollection.IsDirty()`—i.e., if any relationships have actually changed. If they haven't, there's no need to add this extra overhead to `Save()`.

It's worth noting that this combines with a previous condition to skip this process if a) there are any unresolved references (i.e., related topics that aren't yet saved), _and_ this is part of a recursive `Save()`. In that case, `PersistRelations` would get picked up subsequently as part of the handling of `unresolvedRelationships`.