Bypass `Save()` if nothing has changed
Previously, when calling `Save()`, the `UpdateTopic` stored procedure would be called even if no data was changed. This is because the `Save()` method didn't have the means to determine if anything had changed.

With the introduction of `AttributeValueCollection.IsDirty()` (d3b49ff) and `RelatedTopicCollection.IsDirty()` (2adef58), we got two steps closer to being able to do that.

In addition, however, we also need to check for mismatches between indexed and extended attributes. This is easy with indexed attributes, as we only need to check to see if there are any indexed attributes returned from `GetAttributes()` (6662f07). Since all extended attributes are returned, regardless of whether they were changed, we can do this by looking for attributes whose `IsExtendedAttribute` property is set to `false`, indicating the configuration has changed since they were last saved (4762694).

This allows us to detect not only whether any attributes or relationships have changed, but _also_ whether the _storage location_ for any attributes has changed. If none of those are true, then we can safely bypass the `UpdateTopic` stored procedure entirely.

This is valuable for recursive `Save()`s using the `SqlTopicRepository`, including those triggered by the **OnTopic Data Transfer** library's `Topic.Import()`, as there's a lot of overhead to saving every topic, even if it hasn't otherwise changed. There's still quite a bit of overhead in evaluating whether or not anything has changed, but that dwarfs in comparison to actually calling the database.

(To put some rough numbers to this, running a recursive `Save()` on the current reference topic graph for `Root:Configuration` took over two minutes without this, which each `Save()` adding between 55-125ms. With this in place, and no changes to topics, it takes one second. Obviously, it'll take longer when data needs to be persisted, but the point is we can reduce that considerably if we are confident that no data has changed.)