Move variables to top of `Save()`
This not only makes it easier to get a sense of `Save()'s scope of responsibility, but more importantly will allow us to introduce an escape clause for `isDirty` in a future commit, thus preventing the need to do subsequent processing if `!isDirty`.

As part of this, renamed `excludeLastModified` to `areAttributesDirty`, which is more descriptive given that it's no longer being used exclusively for tracking `excludeLastModified`. This is more consistent with `areRelationshipsDirty` as well.