Implement the new `AttributeValuesDataTable` class
In a previous commit (313fe71) we introduced a new `AttributeValuesDataTable` which centralizes the definition of the `AttributeValues` table-valued type in the SQL Server database schema. In this commit, I implement that in the `SqlTopicRepository`, simplifying the code for `Save()`.

As part of this, renamed `attributes` to `attributeValues` for consistency with the data table and type name.