Implement the new `TopicListDataTable` class
In a previous commit (7a37dc9) we introduced a new `TopicListDataTable` which centralizes the definition of the `TopicList` table-valued type in the SQL Server database schema. In this commit, I implement that in the `SqlTopicRepository`, simplifying the code for `PersistRelations()`.

As part of this, I moved the construction of the data table into the `for` loop so that we don't need to `Clear()` it each iteration. Not sure where this nets out in terms of performance, but it makes for cleaner and more intuitive code.