Removed legacy `CreateRelationshipsXml()` support
In a previous version of OnTopic, presumably, there was a need to build an XML document out of the relationships, presumably so that they could optionally be stored as part of the XML blob. This hasn't been used for a while, however, and only adds clutter and unnecessary complexity to an already bloated class. Cancelled!