Share single `version` across recursive `Save()`
Previously, `version` was defined on each iteration of `Save()`. This means if a recursive `Save()` occurred, each topic would have a _slightly_ different `Version` in the database. This didn't hurt anything, but could make it difficult to track what `Save()`s occurred as part of a single request when debugging.

More importantly, with the unresolved reference handling (0c860ae), this meant that the same topic could end up with two different versions generated for the same `Save()`.

This remedies both issues by centralizing the definition via the public `Save()` entry point.