Removed legacy suppressions from Microsoft Code Contracts
Microsoft Code Contract's static code analysis was prone to a number of false positives, which necessitated a few `[SuppressMessage()]` attributes to be sprinkled in the code. Since we no longer use Microsoft Code Contracts—and, in fact, the static code analysis isn't even supported with .NET Standard!—we're removing these suppressions.