Migrated to C# 8's new inline `using` syntax
With C# 8, we no longer need to nest `IDisposable`s inside a `using {}` block or manually call `Dispose()` as part of e.g. a `finally {}` block. Instead, by prefixing the `IDisposable`'s declaration with `using`, they will automatically be disposed of when the code execution exits the current block. This allows us to entirely remove all of the `finally {}` blocks.

Because `using var variable` doesn't align neatly with `var variable`, I've added a line break between the declaration of `using var` and `var` blocks. I'm not sure if this is a best practice yet—it's a gap in our style guide for now—but I _think_ it's a preference. We'll keep evaluating this as we implement the pattern more frequently.